Create patient endpoints

+++ src/HospitalManagementSystem.Api/Repositories/PatientsRepository.cs
@@ -19,13 +19,17 @@ namespace HospitalManagementSystem.Api.Repositories
 
         public async Task DeletePatient(string patientId)
         {
+            if (string.IsNullOrWhiteSpace(patientId)) { throw new ArgumentNullException(nameof(patientId)); }
+
             var patientIdFilter = Builders<PatientReadModel>.Filter.Eq(x => x._id, patientId);
 
             await _db.GetCollection<PatientReadModel>(typeof(PatientReadModel).Name).DeleteOneAsync(patientIdFilter);
         }
 
-        public async Task<PatientReadModel> GetPatientById(string patientId)
+        public async Task<PatientReadModel?> GetPatientById(string patientId)
         {
+            if (string.IsNullOrWhiteSpace(patientId)) { throw new ArgumentNullException(nameof(patientId)); }
+
             var filter = Builders<PatientReadModel>.Filter.Eq(x => x._id, patientId);
 
             return (await _db.GetCollection<PatientReadModel>(typeof(PatientReadModel).Name).FindAsync(filter)).FirstOrDefault();
@@ -45,7 +49,6 @@ namespace HospitalManagementSystem.Api.Repositories
             filters.AddFilter((x) => x.IsAdmitted, query.isAdmitted);
             filters.AddFilter((x) => x.PatientStatus, query.Status);
             filters.AddFilter((x) => x.Gender, query.Gender);
-
 
             var filter = filters.Consolidate();
 
@@ -79,18 +82,48 @@ namespace HospitalManagementSystem.Api.Repositories
 
         public async Task UpsertPatient(PatientReadModel cmd)
         {
+            if (cmd == null) throw new ArgumentNullException(nameof(cmd));
+
             var filter = Builders<PatientReadModel>.Filter.Eq(x => x._id, cmd._id);
-            var update = Builders<PatientReadModel>.Update
-                .Set(x => x._id, cmd._id)
-                .Set(x => x.Name, cmd.Name)
-                .Set(x => x.DateOfBirth, cmd.DateOfBirth)
-                .Set(x => x.Gender, cmd.Gender)
-                .Set(x => x.PhoneNumber, cmd.PhoneNumber)
-                .Set(x => x.Email, cmd.Email)
-                .Set(x => x.AdmissionDate, cmd.AdmissionDate)
-                .Set(x => x.IsAdmitted, cmd.IsAdmitted)
-                .Set(x => x.RoomId, cmd.RoomId)
-                .Set(x => x.PatientStatus, cmd.PatientStatus);
+            var update = Builders<PatientReadModel>.Update.Set(x => x._id, cmd._id);
+
+            if (!string.IsNullOrWhiteSpace(cmd._id))
+            {
+                update = update.Set(x => x._id, cmd._id);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.Name))
+            {
+                update = update.Set(x => x.Name, cmd.Name);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.DateOfBirth))
+            {
+                update = update.Set(x => x.DateOfBirth, cmd.DateOfBirth);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.Gender))
+            {
+                update = update.Set(x => x.Gender, cmd.Gender);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.PhoneNumber))
+            {
+                update = update.Set(x => x.PhoneNumber, cmd.PhoneNumber);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.Email))
+            {
+                update = update.Set(x => x.Email, cmd.Email);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.PatientStatus))
+            {
+                update = update.Set(x => x.PatientStatus, cmd.PatientStatus);
+            }
+            if (!string.IsNullOrWhiteSpace(cmd.AdmissionDate))
+            {
+                update = update.Set(x => x.AdmissionDate, cmd.AdmissionDate);
+            }
+            if (cmd.RoomId.HasValue)
+            {
+                update = update.Set(x => x.RoomId, cmd.RoomId);
+            }
+            update = update.Set(x => x.IsAdmitted, cmd.IsAdmitted);
 
             var options = new UpdateOptions { IsUpsert = true };
